refactor: put files regarding dashboard to "admin" folder
