+++ client/components/page/Admin/Dashboard/Dashboard.page.tsx
@@ -1,13 +1,13 @@
 import { Footer } from '@/components/shared/Footer';
 import { SidebarWithHeader } from '@/components/shared/Header';
-import { DashboardNextPageProps } from 'pages/dashboard';
-import Dashoard from './Dashboard';
+import { DashboardNextPageProps } from 'pages/admin/dashboard';
+import Dashboard from './Dashboard';
 
 const DashboardPage = (props: DashboardNextPageProps) => {
   return (
     <>
       <SidebarWithHeader>
-        <Dashoard {...props} />
+        <Dashboard {...props} />
       </SidebarWithHeader>
       <Footer />
     </>
