--- client/components/page/Dashboard/index.ts
@@ -1,1 +1,0 @@
-export { default as DashboardPage } from './Dashboard.page';
