+++ client/components/page/List/List.tsx
@@ -1,13 +1,8 @@
-import { SimpleTable } from '@/components/shared/Table';
 import { ListNextPageProps } from 'pages/list';
 
 const List = (props: ListNextPageProps) => {
   console.log(props);
-  return (
-    <main>
-      <SimpleTable />
-    </main>
-  );
+  return <main></main>;
 };
 
 export default List;
