+++ client/components/shared/Icon/WalletIcon.tsx
@@ -1,0 +1,20 @@
+import { createIcon } from '@chakra-ui/icons';
+
+const WalletIcon = createIcon({
+  displayName: 'WalletIcon',
+  viewBox: '0 0 24 24',
+  path: (
+    <g>
+      <path
+        fill="currentColor"
+        d="M4.447 4.818h14.062c.164 0 .328.01.491.031a2.9 2.9 0 00-3.406-2.441L4.03 4.382h-.013a2.9 2.9 0 00-1.805 1.149 3.848 3.848 0 012.236-.713zM18.51 5.875H4.446a2.816 2.816 0 00-2.813 2.812v8.438a2.816 2.816 0 002.813 2.812h14.062a2.815 2.815 0 002.813-2.812V8.687a2.815 2.815 0 00-2.813-2.812zm-2.088 8.437a1.406 1.406 0 110-2.811 1.406 1.406 0 010 2.811z"
+      />
+      <path
+        fill="currentColor"
+        d="M1.656 11.651V7.28c0-.952.528-2.549 2.358-2.895 1.553-.291 3.091-.291 3.091-.291s1.011.703.176.703-.813 1.077 0 1.077 0 1.032 0 1.032L4.007 10.62l-2.35 1.032z"
+      />
+    </g>
+  ),
+});
+
+export default WalletIcon;
