+++ client/components/shared/Icon/index.ts
@@ -1,1 +1,2 @@
 export { default as StatsIcon } from './StatsIcon';
+export { default as WalletIcon } from './WalletIcon';
